fix: move some interfaces in frontend
