+++ frontend/src/interfaces/statistics.interface.ts
@@ -39,7 +39,7 @@ export interface Message {
   attachments: Attachment[];
 }
 
-interface PagingMetaData {
+export interface PagingMetaData {
   page: number;
   limit: number;
   count: number;
@@ -74,35 +74,10 @@ export interface Status {
   unsuccessful: number;
 }
 
-export interface UserRecLetters {
-  _meta: PagingMetaData;
-  letters: RecLetter[];
-}
-
-export interface RecLetter {
-  id: string;
-  subject: string;
-  municipalityId: string;
-  status: string;
-  body: string;
-  contentType: string;
-  created: string;
-  updated: string;
-  supportInfo: RecSupportInfo;
-  attachments: RecAttachment[];
-}
-
 export interface RecAttachment {
   id: string;
   fileName: string;
   contentType: string;
-}
-
-export interface RecSupportInfo {
-  supportText: string;
-  contactInformationUrl: string;
-  contactInformationPhoneNumber: string;
-  contactInformationEmail: string;
 }
 
 export interface SigningInfo {
